Move method classifications to separate component
